requirements.txt
flask
pywin32
Pillow
requests